refactor: move lib out of specs
