+++ tests/schema/specs/settings.spec.ts
@@ -1,7 +1,7 @@
 import gql from 'noop-tag';
 import { expect, test } from 'vitest';
 
-import { fetch } from './lib';
+import { fetch } from '../lib';
 
 test('Page', async () => {
   const result = await fetch(gql`
